refactor(frontend): split shortform wizard utils
